+++ src/app/pages/homepage/page.tsx
@@ -1,6 +1,6 @@
 'use client';
 import React, { useState } from 'react';
-import Post from '@/app/context/postContext';
+import Post from '@/app/components/post/Post';
 
 interface PostData {
   userid: string;
@@ -16,22 +16,22 @@ interface PostData {
 const HomePage: React.FC = () => {
   const [posts, setPosts] = useState<PostData[]>([
     {
-      userid: 'USER-0',
-      title: 'Post Title 1',
-      description: 'DESCRIPTION 1',
-      media: 'https://via.placeholder.com/150', 
+      userid: 'MarcuSV0',
+      title: 'El patio de mi casa',
+      description: 'Quien quiere venir a mi casa? ',
+      media: 'https://1.bp.blogspot.com/-zKX8CREi3QY/T2OMZgW3s6I/AAAAAAAAWzg/5sFH754c6sw/s1600/Los-mas-Hermosos-Paisajes-Naturales_04.jpg', 
       likes: 20,
       updateDate: new Date(),
       comments: 5,
       favorites: 10,
     },
     {
-      userid: 'USER-1',
-      title: 'Post Title 2',
-      description: 'DESCRIPTION 2',
-      media: 'https://via.placeholder.com/150', 
-      likes: 15,
-      updateDate: new Date(),
+      userid: 'DANIEL-1',
+      title: 'Mi viaje a Europa',
+      description: 'Nest way to find my proposit',
+      media: 'https://th.bing.com/th/id/OIF.0mpwGSIBRUhbxBmKJHCrpA?rs=1&pid=ImgDetMain', 
+      likes: 1,
+      updateDate: new Date(23/3/23),
       comments: 3,
       favorites: 8,
     },
